places-sidebar: allow dropping of files to bookmarks
Change the way DnD works in the places sidebar.
After this commit, DnD on the bookmarks area will always mean
"copy/move the DnD source into the bookmark" instead of "add a bookmark
to this location" (DnD of bookmark rows in their section will still
reorder them), since that's the most common action you want to do.
This also makes it consistent with the default action obtained by
dropping on the built-in places below.

https://bugzilla.gnome.org/show_bug.cgi?id=645353