toolbar: ensure square buttons on the toolbar
Since we use symbolic 16px icons inside tool buttons, we should make
sure they still get a square aspect, and buttons of the right size.

As we create the contents of our tool buttons manually, ensure this
by setting a margin on the tool button's GtkImage, sized as the
difference in pixels between the icon we render and the icon size
expected by GtkToolbar.