Change default executable text action to display
Presenting a scary dialog by default isn't nice. The file manager
is not primarily an interface for launch applications. Some filesystems
mark all files as executable and it just sucks to keep asking when you
want to view text files. Also, any application launched this way is
going to suck. You won't have app tracking in the shell and showing
a terminal to users that don't expect it is really nasty.

https://bugzilla.gnome.org/show_bug.cgi?id=601736